Safari-609.1.11.1/LayoutTests/http/tests/security/contentTypeOptions/nosniff-script-without-content-type-blocked.html

<!DOCTYPE html>
<html>
<head>
    <title>'X-Content-Type-Options: nosniff;' blocks scripts!</title>
<body>
    <script src="/js-test-resources/js-test-pre.js"></script>
    <script>
        description('Check that script sent with an \'X-Content-Type-Options: nosniff\' header is correctly blocked if no \'Content-Type\' header is present.');
        window.jsTestIsAsync = true;
        window.scriptsSuccessfullyLoaded = 0;
        window.onload = function () {
            shouldBe('window.scriptsSuccessfullyLoaded', '0');
            finishJSTest();
        };
    </script>
    <script src="./resources/script-with-header.pl"></script>
    <script src="/js-test-resources/js-test-post.js"></script>
</body>
</html>
